fixed channels height bug
The channels holder was longer than it should be. The new 'display: block' overwrites antiscroll's 'display: inline-block' which broke the flex-box. Now The channels holder is exactly as long as it should be.
Also - the sidebars flex-box didn't work in Firefox. The reason was 'position: absolute' - apparently it disables flex-box functionality in Firefox. Because of that, an additional div (.flexHolder) was added (sucks,  but hey - flex box will get mature and then we can remove this div)