Active Channel Indicator Love
With the recent antiscroll-updates I added a black left-border to the content. As it turned out this black border cut off the active channel indicator which is inside the sidebar but has to be directly connected to the content background for a seamless look.
On the other hand this black border came in quite handy as a separator between personal channel + search and the content.
Solution: a black box-shadow line, a 1 pixel gap between sidebar and content. personal channel and pixel have the same width (233 pixel), the listed channels which can have the active channel indicator have the full width (e.g. bridge the gap) of 234 pixels.