Unread Counter Design Love
Until now, the borders of the circle-shaped counters were blurry. I used a dark translucent border, because the border overlays the background and thus darkens the border but keeps the color hue. Keep in mind, that we want to keep this generic because we have different-colored counters. But the border didn't render well.
So now option two: a box-shadow around it with the color-variables darkened via stylus. Works well.