Playing around with bubbling
Currently using CSS transitions. Works great for single bubbling. The problem is when there are multiple channels bubbling at the same time, animation distances have to be updated on the fly. That doesn't work with CSS transitions. Next try is javascript animations. They might not perform all to good on mobile devices but on desktop they shouldn't make a difference.