discover globe - a test for an out-of-the-box approach to discover new channels
A playful discovery of new channels
======================

The map is a infinite canvas. The users channel is placed in the center with a battle-fog around it. In different directions there are different channels pooled by a certain criteria to explore. For example, in the direction of the upper left corner there are 'channels you might like', based on the channels the user currently follows. In the direction of the upper right corner there are popular channels from his server etc.

The user explores them by steering his channel around:

left-right-top-bottom - move your channel

TODO:
	- connect channels that follow each other with lines
	- add battle-fog to your channel
	- come up with an algorithm for the position of channels that belong together in a cloud
	- fetch channels and position them around the center

Current state: animation to unfold channels around a certain center. Play with it by pressing 'Enter'.