Do not rely on "position: relative" td for channel title
According to the CSS spec, the effect of "position: relative"
on a td element is undefined. Firefox seems to ignore it, which
means that the channel title positioning used in the channel title
("position: absolute" div inside of "position: relative" td) doesn't
work here. Using just normal positioning is safer and changes the
layout only by a few pixels, so use that.