Stripped all code out of topicpost.coffee, as able to loop over all textareas from src/view/channels/index.coffee.
Issue at line 344 of index.coffee (ChannelPostsView) with looping over each textarea and making it autocomplete.

Also concerned that new subscribers will not be pushed to all autocomplete instances so this maybe something to look at.

My next step is to inject a template to the autocomplete plugin for the display of gravatars in the results set.

See https://github.com/lloydwatkin/buddycloud-webclient branch inline-mention.