rm centerBox with old login and register form
sadly this causes older firefox to not store the password in their
vaults, but i admit, i never saw this working well in any browser.

chromium for example stores only passwords in its vault when a form
submit was successful. in their opinion this has to involve a full
page reload. #fail

so, solution for the future is to store the password in localStorage
until browsers come up with a better solution.

(native XMPP in browser? sometimes wishes can come true :3)